Extract Thread into its own component
